Implementa testes restantes e aumenta score de coverage

+++ src/usecase/categoria/categoria.usecase.ts
@@ -24,6 +24,12 @@ export class CategoriaUseCase implements ICategoriaUseCase {
     }
 
     async removerCategoriaPorId(id: string): Promise<void> {
+        const categoriaExiste = await this.categoriaRepository.findById(id);
+
+        if (!categoriaExiste) {
+            throw new ErroNegocio('categoria-nao-cadastrada');
+        }
+
         await this.categoriaRepository.remove(id);
     }
